+++ libraries/base/Data/Functor/Utils.hs
@@ -58,6 +58,10 @@ instance Applicative (StateL s) where
         let (s', f) = kf s
             (s'', v) = kv s'
         in (s'', f v)
+    liftA2 f (StateL kx) (StateL ky) = StateL $ \s ->
+        let (s', x) = kx s
+            (s'', y) = ky s'
+        in (s'', f x y)
 
 -- right-to-left state transformer
 newtype StateR s a = StateR { runStateR :: s -> (s, a) }
@@ -73,6 +77,10 @@ instance Applicative (StateR s) where
         let (s', v) = kv s
             (s'', f) = kf s'
         in (s'', f v)
+    liftA2 f (StateR kx) (StateR ky) = StateR $ \ s ->
+        let (s', y) = ky s
+            (s'', x) = kx s'
+        in (s'', f x y)
 
 -- See Note [Function coercion]
 (#.) :: Coercible b c => (b -> c) -> (a -> b) -> (a -> c)